refactor and standardize codebase
