+++ src/app/properties/[id]/not-found.tsx
@@ -1,5 +1,3 @@
-import React from "react";
-
 export default function PropertyNotFound() {
   return <div>PropertyNotFound</div>;
 }
