+++ src/app/properties/[id]/payments/page.tsx
@@ -1,4 +1,4 @@
-import ComingSoon from "~/components/ui/pages/ComingSoon";
+import { ComingSoon } from "~/components/pages/coming-soon";
 
 export default function Payments() {
   return <ComingSoon />;
