+++ src/app/properties/[id]/settings/page.tsx
@@ -1,4 +1,4 @@
-import ComingSoon from "~/components/ui/pages/ComingSoon";
+import { ComingSoon } from "~/components/pages/coming-soon";
 
 export default function SettingsPage() {
   return <ComingSoon />;
